system/requirements/requirements_textgen.txt
coqui-tts>=0.24.1
faster-whisper>=1.0.*
fuzzywuzzy
gradio>=4.26.0
importlib_metadata>=7.1.*
inputimeout>=1.0.4
Jinja2>=3.1.*
librosa>=0.10.2.post1
local_attention>=1.9.*
nvidia-cublas-cu11>=11.11.3.6
nvidia-cudnn-cu11>=9.1.1.17
onnxruntime-gpu>=1.18.*
pydantic>=2.7.*
python-ffmpeg>=2.0.*
python-Levenshtein>=0.25.1
praat-parselmouth>=0.4.*
pyworld>=0.3.*
sounddevice>=0.4.7
soundfile>=0.12.*
spacy>=3.7.1
torchcrepe>=0.0.2
tqdm>=4.66.*
unidic-lite>=1.0.8
uvicorn>=0.29.0
pypinyin
hangul_romanize
git+https://github.com/huggingface/parler-tts.git
./system/config/fairseq-0.12.4-cp311-cp311-win_amd64.whl; sys_platform == "win32"
./system/config/fairseq-0.12.3.1-cp311-cp311-linux_x86_64.whl; sys_platform == "linux"
piper-tts; sys_platform == "linux"
piper-tts; sys_platform == "darwin"
piper-phonemize==1.1.0; sys_platform == "darwin"